[Product]Feat: 주문 생성 Feign 생성

+++ product/src/main/java/com/logistics/product/application/dto/ProductUpdateReqDto.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 @AllArgsConstructor
 public class ProductUpdateReqDto {
     private String name;
-    private Long stock;
+    private int stock;
     private String description;
     private BigDecimal price;
 }
